+++ pennylane_lightning/core/src/simulators/lightning_tensor/tncuda/gates/tests/Test_MPSTNCuda_Param.cpp
@@ -36,8 +36,7 @@ namespace cuUtil = Pennylane::LightningGPU::Util;
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::PhaseShift", "[MPSTNCuda_Param]", float,
                    double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -45,6 +44,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::PhaseShift", "[MPSTNCuda_Param]", float,
         DevTag<int> dev_tag{0, 0};
 
         const std::vector<TestType> angles{0.3, 0.8, 2.4};
+        const TestType sign = (inverse) ? -1.0 : 1.0;
         const cp_t coef(1.0 / (2 * std::sqrt(2)), 0);
 
         std::vector<std::vector<cp_t>> ps_data;
@@ -82,7 +82,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::PhaseShift", "[MPSTNCuda_Param]", float,
                                       {{0}, {1}, {2}}, {false, false, false});
 
             mps_state.applyOperation("PhaseShift", {index}, inverse,
-                                     {angles[index]});
+                                     {sign * angles[index]});
 
             auto results = mps_state.getDataVector();
 
@@ -92,7 +92,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::PhaseShift", "[MPSTNCuda_Param]", float,
 }
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::RX", "[MPSTNCuda_Param]", float, double) {
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -105,6 +105,10 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::RX", "[MPSTNCuda_Param]", float, double) {
         std::vector<cp_t> results = {{0.34958337, -0.05283436},
                                      {0.32564424, -0.13768018},
                                      {0.12811281, -0.32952558}};
+
+        for (auto &val : results) {
+            val = inverse ? std::conj(val) : val;
+        }
 
         std::vector<std::vector<cp_t>> expected_results{
             std::vector<cp_t>(std::size_t{1} << num_qubits, results[0]),
@@ -130,7 +134,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::RX", "[MPSTNCuda_Param]", float, double) {
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::RY", "[MPSTNCuda_Nonparam]", float,
                    double) {
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -165,6 +169,23 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::RY", "[MPSTNCuda_Nonparam]", float,
                                                          {-0.20141277, 0},
                                                          {0.45763839, 0}}};
 
+        if (inverse) {
+            std::swap(expected_results[0][4], expected_results[0][0]);
+            std::swap(expected_results[0][5], expected_results[0][1]);
+            std::swap(expected_results[0][6], expected_results[0][2]);
+            std::swap(expected_results[0][7], expected_results[0][3]);
+
+            std::swap(expected_results[1][2], expected_results[1][0]);
+            std::swap(expected_results[1][3], expected_results[1][1]);
+            std::swap(expected_results[1][6], expected_results[1][4]);
+            std::swap(expected_results[1][7], expected_results[1][5]);
+
+            std::swap(expected_results[2][1], expected_results[2][0]);
+            std::swap(expected_results[2][3], expected_results[2][2]);
+            std::swap(expected_results[2][5], expected_results[2][4]);
+            std::swap(expected_results[2][7], expected_results[2][6]);
+        }
+
         SECTION("Apply different wire indices") {
             const std::size_t index = GENERATE(0, 1, 2);
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -182,7 +203,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::RY", "[MPSTNCuda_Nonparam]", float,
 }
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::RZ", "[MPSTNCuda_Param]", float, double) {
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -220,6 +241,12 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::RZ", "[MPSTNCuda_Param]", float, double) {
              {0.12811281, -0.32952558},
              {0.12811281, 0.32952558}}};
 
+        for (auto &vec : expected_results) {
+            for (auto &val : vec) {
+                val = inverse ? std::conj(val) : val;
+            }
+        }
+
         SECTION("Apply different wire indices") {
             const std::size_t index = GENERATE(0, 1, 2);
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -238,8 +265,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::RZ", "[MPSTNCuda_Param]", float, double) {
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::ControlledPhaseShift",
                    "[MPSTNCuda_Param]", float, double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -247,6 +273,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::ControlledPhaseShift",
         DevTag<int> dev_tag{0, 0};
 
         const std::vector<TestType> angles{0.3, 2.4};
+        const TestType sign = (inverse) ? -1.0 : 1.0;
         const cp_t coef(1.0 / (2 * std::sqrt(2)), 0);
 
         std::vector<std::vector<cp_t>> ps_data;
@@ -273,7 +300,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::ControlledPhaseShift",
                                       {{0}, {1}, {2}}, {false, false, false});
 
             mps_state.applyOperation("ControlledPhaseShift", {0, 1}, inverse,
-                                     {angles[0]});
+                                     {sign * angles[0]});
 
             auto results = mps_state.getDataVector();
 
@@ -287,7 +314,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::ControlledPhaseShift",
                                       {{0}, {1}, {2}}, {false, false, false});
 
             mps_state.applyOperation("ControlledPhaseShift", {0, 2}, inverse,
-                                     {angles[1]});
+                                     {sign * angles[1]});
 
             auto results = mps_state.getDataVector();
 
@@ -298,8 +325,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::ControlledPhaseShift",
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::Rot", "[MPSTNCuda_param]", float,
                    double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -320,8 +346,10 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::Rot", "[MPSTNCuda_param]", float,
             const auto rot_mat =
                 Pennylane::Gates::getRot<std::complex, TestType>(
                     angles[i][0], angles[i][1], angles[i][2]);
-            expected_results[i][0] = rot_mat[0];
-            expected_results[i][0b1 << (num_qubits - i - 1)] = rot_mat[2];
+            expected_results[i][0] =
+                inverse ? std::conj(rot_mat[0]) : rot_mat[0];
+            expected_results[i][0b1 << (num_qubits - i - 1)] =
+                inverse ? -rot_mat[2] : rot_mat[2];
         }
 
         SECTION("Apply at different wire indices") {
@@ -337,8 +365,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::Rot", "[MPSTNCuda_param]", float,
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRot", "[MPSTNCuda_param]", float,
                    double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -375,8 +402,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRot", "[MPSTNCuda_param]", float,
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingXX", "[MPSTNCuda_param]", float,
                    double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -391,6 +417,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingXX", "[MPSTNCuda_param]", float,
             std::vector<cp_t>(1 << num_qubits),
             std::vector<cp_t>(1 << num_qubits),
             std::vector<cp_t>(1 << num_qubits)};
+
         expected_results[0][0] = {0.9887710779360422, 0.0};
         expected_results[0][6] = {0.0, -0.14943813247359922};
 
@@ -402,6 +429,12 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingXX", "[MPSTNCuda_param]", float,
 
         expected_results[3][0] = {0.9210609940028851, 0.0};
         expected_results[3][5] = {0.0, -0.3894183423086505};
+
+        for (auto &vec : expected_results) {
+            for (auto &val : vec) {
+                val = inverse ? std::conj(val) : val;
+            }
+        }
 
         SECTION("Apply adjacent wires") {
             const std::size_t index = GENERATE(0, 1);
@@ -430,8 +463,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingXX", "[MPSTNCuda_param]", float,
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingXY", "[MPSTNCuda_param]", float,
                    double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -455,6 +487,12 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingXY", "[MPSTNCuda_param]", float,
         expected_results[1][3] = {0.34958337, 0.05283436};
         expected_results[1][4] = {0.34958337, 0.05283436};
         expected_results[1][6] = {0.34958337, 0.05283436};
+
+        for (auto &vec : expected_results) {
+            for (auto &val : vec) {
+                val = inverse ? std::conj(val) : val;
+            }
+        }
 
         SECTION("Apply adjacent wires") {
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -486,8 +524,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingXY", "[MPSTNCuda_param]", float,
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingYY", "[MPSTNCuda_param]", float,
                    double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -511,6 +548,12 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingYY", "[MPSTNCuda_param]", float,
         expected_results[1][3] = {0.34958337, -0.05283436};
         expected_results[1][4] = {0.34958337, -0.05283436};
         expected_results[1][6] = {0.34958337, -0.05283436};
+
+        for (auto &vec : expected_results) {
+            for (auto &val : vec) {
+                val = inverse ? std::conj(val) : val;
+            }
+        }
 
         SECTION("Apply adjacent wires") {
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -542,8 +585,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingYY", "[MPSTNCuda_param]", float,
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingZZ", "[MPSTNCuda_param]", float,
                    double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -567,6 +609,12 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingZZ", "[MPSTNCuda_param]", float,
         expected_results[1][2] = {0.34958337, -0.05283436};
         expected_results[1][5] = {0.34958337, -0.05283436};
         expected_results[1][7] = {0.34958337, -0.05283436};
+
+        for (auto &vec : expected_results) {
+            for (auto &val : vec) {
+                val = inverse ? std::conj(val) : val;
+            }
+        }
 
         SECTION("Apply adjacent wires") {
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -598,8 +646,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::IsingZZ", "[MPSTNCuda_param]", float,
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRX", "[MPSTNCuda_param]", float,
                    double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -623,6 +670,12 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRX", "[MPSTNCuda_param]", float,
         expected_results[1][5] = {0.34958337, -0.05283436};
         expected_results[1][6] = {0.34958337, -0.05283436};
         expected_results[1][7] = {0.34958337, -0.05283436};
+
+        for (auto &vec : expected_results) {
+            for (auto &val : vec) {
+                val = inverse ? std::conj(val) : val;
+            }
+        }
 
         SECTION("Apply adjacent wires") {
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -655,8 +708,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRX", "[MPSTNCuda_param]", float,
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRY", "[MPSTNCuda_param]", float,
                    double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -676,10 +728,20 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRY", "[MPSTNCuda_param]", float,
         expected_results[0][6] = {0.40241773, 0.0};
         expected_results[0][7] = {0.40241773, 0.0};
 
+        if (inverse) {
+            std::swap(expected_results[0][4], expected_results[0][6]);
+            std::swap(expected_results[0][5], expected_results[0][7]);
+        }
+
         expected_results[1][4] = {0.29674901, 0.0};
         expected_results[1][5] = {0.40241773, 0.0};
         expected_results[1][6] = {0.29674901, 0.0};
         expected_results[1][7] = {0.40241773, 0.0};
+
+        if (inverse) {
+            std::swap(expected_results[1][4], expected_results[1][5]);
+            std::swap(expected_results[1][6], expected_results[1][7]);
+        }
 
         SECTION("Apply adjacent wires") {
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -712,8 +774,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRY", "[MPSTNCuda_param]", float,
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRZ", "[MPSTNCuda_param]", float,
                    double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -737,6 +798,12 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRZ", "[MPSTNCuda_param]", float,
         expected_results[1][5] = {0.34958337, 0.05283436};
         expected_results[1][6] = {0.34958337, -0.05283436};
         expected_results[1][7] = {0.34958337, 0.05283436};
+
+        for (auto &vec : expected_results) {
+            for (auto &val : vec) {
+                val = inverse ? std::conj(val) : val;
+            }
+        }
 
         SECTION("Apply adjacent wires") {
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -768,8 +835,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::CRZ", "[MPSTNCuda_param]", float,
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitation", "[MPSTNCuda_param]",
                    float, double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -793,6 +859,13 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitation", "[MPSTNCuda_param]",
         expected_results[1][3] = {0.29674901, 0.0};
         expected_results[1][4] = {0.40241773, 0.0};
         expected_results[1][6] = {0.40241773, 0.0};
+
+        if (inverse) {
+            std::swap(expected_results[0][2], expected_results[0][4]);
+            std::swap(expected_results[0][3], expected_results[0][5]);
+            std::swap(expected_results[1][1], expected_results[1][6]);
+            std::swap(expected_results[1][3], expected_results[1][4]);
+        }
 
         SECTION("Apply adjacent wires") {
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -827,8 +900,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitation", "[MPSTNCuda_param]",
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitationMinus",
                    "[MPSTNCuda_param]", float, double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -843,6 +915,12 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitationMinus",
             std::vector<cp_t>(1 << num_qubits, {0.34958337, -0.05283436}),
         };
 
+        for (auto &vec : expected_results) {
+            for (auto &val : vec) {
+                val = inverse ? std::conj(val) : val;
+            }
+        }
+
         expected_results[0][2] = {0.29674901, 0.0};
         expected_results[0][3] = {0.29674901, 0.0};
         expected_results[0][4] = {0.40241773, 0.0};
@@ -852,6 +930,13 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitationMinus",
         expected_results[1][3] = {0.29674901, 0.0};
         expected_results[1][4] = {0.40241773, 0.0};
         expected_results[1][6] = {0.40241773, 0.0};
+
+        if (inverse) {
+            std::swap(expected_results[0][2], expected_results[0][4]);
+            std::swap(expected_results[0][3], expected_results[0][5]);
+            std::swap(expected_results[1][1], expected_results[1][6]);
+            std::swap(expected_results[1][3], expected_results[1][4]);
+        }
 
         SECTION("Apply adjacent wires") {
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -886,8 +971,7 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitationMinus",
 
 TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitationPlus",
                    "[MPSTNCuda_param]", float, double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
         using cp_t = std::complex<TestType>;
         std::size_t num_qubits = 3;
@@ -902,6 +986,12 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitationPlus",
             std::vector<cp_t>(1 << num_qubits, {0.34958337, 0.05283436}),
         };
 
+        for (auto &vec : expected_results) {
+            for (auto &val : vec) {
+                val = inverse ? std::conj(val) : val;
+            }
+        }
+
         expected_results[0][2] = {0.29674901, 0.0};
         expected_results[0][3] = {0.29674901, 0.0};
         expected_results[0][4] = {0.40241773, 0.0};
@@ -911,6 +1001,13 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitationPlus",
         expected_results[1][3] = {0.29674901, 0.0};
         expected_results[1][4] = {0.40241773, 0.0};
         expected_results[1][6] = {0.40241773, 0.0};
+
+        if (inverse) {
+            std::swap(expected_results[0][2], expected_results[0][4]);
+            std::swap(expected_results[0][3], expected_results[0][5]);
+            std::swap(expected_results[1][1], expected_results[1][6]);
+            std::swap(expected_results[1][3], expected_results[1][4]);
+        }
 
         SECTION("Apply adjacent wires") {
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
@@ -942,236 +1039,39 @@ TEMPLATE_TEST_CASE("MPSTNCuda::Gates::SingleExcitationPlus",
     }
 }
 
-TEMPLATE_TEST_CASE("MPSTNCuda::Gates::DoubleExcitation", "[MPSTNCuda_param]",
+TEMPLATE_TEST_CASE("MPSTNCuda::Param_Gates::2+_wires", "[MPSTNCuda_Param]",
                    float, double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
+    const bool inverse = GENERATE(false, true);
     {
-        using cp_t = std::complex<TestType>;
-        std::size_t num_qubits = 5;
         std::size_t maxExtent = 2;
         DevTag<int> dev_tag{0, 0};
 
-        const std::vector<TestType> angles = {0.3};
+        SECTION("DoubleExcitation gate") {
+            std::size_t num_qubits = 5;
 
-        // Results collected from `default.qubit`
-        std::vector<std::vector<cp_t>> expected_results{
-            std::vector<cp_t>(1 << num_qubits, {0.1767767, 0.0}),
-            std::vector<cp_t>(1 << num_qubits, {0.1767767, 0.0}),
-        };
-
-        expected_results[0][6] = {0.1483745, 0.0};
-        expected_results[0][7] = {0.1483745, 0.0};
-        expected_results[0][24] = {0.20120886, 0.0};
-        expected_results[0][25] = {0.20120886, 0.0};
-
-        expected_results[1][5] = {0.1483745, 0.0};
-        expected_results[1][7] = {0.1483745, 0.0};
-        expected_results[1][24] = {0.20120886, 0.0};
-        expected_results[1][26] = {0.20120886, 0.0};
-
-        SECTION("Apply adjacent wires") {
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
-            mps_state.reset();
 
-            mps_state.applyOperations(
-                {"Hadamard", "Hadamard", "Hadamard", "Hadamard", "Hadamard"},
-                {{0}, {1}, {2}, {3}, {4}}, {false, false, false, false, false});
-
-            mps_state.applyOperation("DoubleExcitation", {0, 1, 2, 3}, inverse,
-                                     angles);
-
-            auto results = mps_state.getDataVector();
-
-            CHECK(results == Pennylane::Util::approx(expected_results[0]));
+            REQUIRE_THROWS_AS(mps_state.applyOperation("DoubleExcitation",
+                                                       {0, 1, 2, 3}, inverse),
+                              LightningException);
         }
 
-        SECTION("Apply non-adjacent wires") {
+        SECTION("DoubleExcitationMinus gate") {
+            std::size_t num_qubits = 5;
+
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
-
-            mps_state.applyOperations(
-                {"Hadamard", "Hadamard", "Hadamard", "Hadamard", "Hadamard"},
-                {{0}, {1}, {2}, {3}, {4}}, {false, false, false, false, false});
-
-            mps_state.applyOperation("DoubleExcitation", {0, 1, 2, 4}, inverse,
-                                     angles);
-
-            auto results = mps_state.getDataVector();
-
-            CHECK(results == Pennylane::Util::approx(expected_results[1]));
-        }
-    }
-}
-
-TEMPLATE_TEST_CASE("MPSTNCuda::Gates::DoubleExcitationMinus",
-                   "[MPSTNCuda_param]", float, double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
-    {
-        using cp_t = std::complex<TestType>;
-        std::size_t num_qubits = 5;
-        std::size_t maxExtent = 2;
-        DevTag<int> dev_tag{0, 0};
-
-        const std::vector<TestType> angles = {0.3};
-
-        // Results collected from `default.qubit`
-        std::vector<std::vector<cp_t>> expected_results{
-            std::vector<cp_t>(1 << num_qubits, {0.17479168, -0.02641718}),
-            std::vector<cp_t>(1 << num_qubits, {0.17479168, -0.02641718}),
-        };
-
-        expected_results[0][6] = {0.1483745, 0.0};
-        expected_results[0][7] = {0.1483745, 0.0};
-        expected_results[0][24] = {0.20120886, 0.0};
-        expected_results[0][25] = {0.20120886, 0.0};
-
-        expected_results[1][5] = {0.1483745, 0.0};
-        expected_results[1][7] = {0.1483745, 0.0};
-        expected_results[1][24] = {0.20120886, 0.0};
-        expected_results[1][26] = {0.20120886, 0.0};
-
-        SECTION("Apply adjacent wires") {
-            MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
-            mps_state.reset();
-
-            mps_state.applyOperations(
-                {"Hadamard", "Hadamard", "Hadamard", "Hadamard", "Hadamard"},
-                {{0}, {1}, {2}, {3}, {4}}, {false, false, false, false, false});
-
-            mps_state.applyOperation("DoubleExcitationMinus", {0, 1, 2, 3},
-                                     inverse, angles);
-
-            auto results = mps_state.getDataVector();
-
-            CHECK(results == Pennylane::Util::approx(expected_results[0]));
+            REQUIRE_THROWS_AS(mps_state.applyOperation("DoubleExcitationMinus",
+                                                       {0, 1, 2, 3}, inverse),
+                              LightningException);
         }
 
-        SECTION("Apply non-adjacent wires") {
+        SECTION("DoubleExcitationPlus gate") {
+            std::size_t num_qubits = 5;
+
             MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
-
-            mps_state.applyOperations(
-                {"Hadamard", "Hadamard", "Hadamard", "Hadamard", "Hadamard"},
-                {{0}, {1}, {2}, {3}, {4}}, {false, false, false, false, false});
-
-            mps_state.applyOperation("DoubleExcitationMinus", {0, 1, 2, 4},
-                                     inverse, angles);
-
-            auto results = mps_state.getDataVector();
-
-            CHECK(results == Pennylane::Util::approx(expected_results[1]));
-        }
-    }
-}
-
-TEMPLATE_TEST_CASE("MPSTNCuda::Gates::DoubleExcitationPlus",
-                   "[MPSTNCuda_param]", float, double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
-    {
-        using cp_t = std::complex<TestType>;
-        std::size_t num_qubits = 5;
-        std::size_t maxExtent = 2;
-        DevTag<int> dev_tag{0, 0};
-
-        const std::vector<TestType> angles = {0.3};
-
-        // Results collected from `default.qubit`
-        std::vector<std::vector<cp_t>> expected_results{
-            std::vector<cp_t>(1 << num_qubits, {0.17479168, 0.02641718}),
-            std::vector<cp_t>(1 << num_qubits, {0.17479168, 0.02641718}),
-        };
-
-        expected_results[0][6] = {0.1483745, 0.0};
-        expected_results[0][7] = {0.1483745, 0.0};
-        expected_results[0][24] = {0.20120886, 0.0};
-        expected_results[0][25] = {0.20120886, 0.0};
-
-        expected_results[1][5] = {0.1483745, 0.0};
-        expected_results[1][7] = {0.1483745, 0.0};
-        expected_results[1][24] = {0.20120886, 0.0};
-        expected_results[1][26] = {0.20120886, 0.0};
-
-        SECTION("Apply adjacent wires") {
-            MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
-            mps_state.reset();
-
-            mps_state.applyOperations(
-                {"Hadamard", "Hadamard", "Hadamard", "Hadamard", "Hadamard"},
-                {{0}, {1}, {2}, {3}, {4}}, {false, false, false, false, false});
-
-            mps_state.applyOperation("DoubleExcitationPlus", {0, 1, 2, 3},
-                                     inverse, angles);
-
-            auto results = mps_state.getDataVector();
-
-            CHECK(results == Pennylane::Util::approx(expected_results[0]));
-        }
-
-        SECTION("Apply non-adjacent wires") {
-            MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
-
-            mps_state.applyOperations(
-                {"Hadamard", "Hadamard", "Hadamard", "Hadamard", "Hadamard"},
-                {{0}, {1}, {2}, {3}, {4}}, {false, false, false, false, false},
-                {{0.0}, {0.0}, {0.0}, {0.0}, {0.0}});
-
-            mps_state.applyOperation("DoubleExcitationPlus", {0, 1, 2, 4},
-                                     inverse, angles);
-
-            auto results = mps_state.getDataVector();
-
-            CHECK(results == Pennylane::Util::approx(expected_results[1]));
-        }
-    }
-}
-
-TEMPLATE_TEST_CASE("MPSTNCuda::Gates::MultiRZ", "[MPSTNCuda_param]", float,
-                   double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
-    {
-        std::size_t num_qubits = 5;
-        std::size_t maxExtent = 2;
-        DevTag<int> dev_tag{0, 0};
-
-        const std::vector<TestType> angles = {0.3};
-        SECTION("Throw errors") {
-            MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
-
-            REQUIRE_THROWS_WITH(mps_state.applyOperation("MultiRZ", {0, 1},
-                                                         inverse, {angles[0]}),
-                                Catch::Contains("Unsupported gate"));
-        }
-    }
-}
-
-TEMPLATE_TEST_CASE("MPSTNCuda::Gates::Matrix", "[MPSTNCuda_param]", float,
-                   double) {
-    // TODO only support inverse = false now
-    const bool inverse = GENERATE(false);
-    {
-        using cp_t = std::complex<TestType>;
-        std::size_t num_qubits = 3;
-        std::size_t maxExtent = 2;
-        DevTag<int> dev_tag{0, 0};
-
-        const std::vector<cp_t> x_gate{cuUtil::ZERO<cp_t>(),
-                                       cuUtil::ONE<cp_t>(), cuUtil::ONE<cp_t>(),
-                                       cuUtil::ZERO<cp_t>()};
-
-        SECTION("Append tensor operator with host data and wires") {
-            const std::size_t index = GENERATE(0, 1, 2);
-            MPSTNCuda<TestType> mps_state{num_qubits, maxExtent, dev_tag};
-
-            mps_state.applyOperation("applyMatrix", {index}, inverse, {},
-                                     x_gate);
-
-            auto results = mps_state.getDataVector();
-
-            CHECK(results[0] == cuUtil::ZERO<std::complex<TestType>>());
-            CHECK(results[0b1 << (num_qubits - index - 1)] ==
-                  cuUtil::ONE<std::complex<TestType>>());
+            REQUIRE_THROWS_AS(mps_state.applyOperation("DoubleExcitationPlus",
+                                                       {0, 1, 2, 3}, inverse),
+                              LightningException);
         }
     }
 }